src/requirements.txt
matplotlib
numpy
pytorch
opencv-python
tensorboardX
flask 
seaborn 
pickle